subclass Replay into OnlineReplay and LocalReplay
