+++ osu-ac/replay.py
@@ -1,35 +1,5 @@
-import base64
-import requests
-
 import numpy as np
 import osrparse
-
-from loader import Loader
-from cacher import Cacher
-
-
-def check_cache(function):
-    """
-    Decorator that checks if the replay by the given user_id on the given map_id is already cached.
-    If so, returns a Replay instance from the cached string instead of requesting it from the api.
-
-    Note that map_id and user_id must be the first and second arguments to the function respectively.
-
-    Returns:
-        A Replay instance from the cached replay if it was cached, or the return value of the function if not.
-    """
-
-    def wrapper(*args, **kwargs):
-        map_id = args[0]
-        user_id = args[1]
-        lzma = Cacher.check_cache(map_id, user_id)
-        if(lzma):
-            replay_data = osrparse.parse_replay(lzma, pure_lzma=True).play_data
-            return Replay(replay_data, user_id)
-        else:
-            function(*args, **kwargs)
-    return wrapper
-
 
 class Interpolation:
     """A utility class containing coordinate interpolations."""
@@ -70,7 +40,6 @@ class Replay:
         String player_name: The player who set the replay. Often given as a player id.
     """
 
-
     def __init__(self, replay_data, player_name):
         """
         Initializes a Replay instance.
@@ -83,49 +52,6 @@ class Replay:
 
         self.player_name = player_name
         self.play_data = replay_data
-
-    @staticmethod
-    @check_cache
-    def from_map(map_id, user_id, cache):
-        """
-        Creates a Replay instance from a replay by the given user on the given map.
-
-        Args:
-            String map_id: The map_id to download the replay from.
-            String user_id: The user id to download the replay of.
-                            Also used as the username of the Replay.
-
-        Returns:
-            The Replay instance created with the given information.
-        """
-
-        replay_data_string = Loader.replay_data(map_id, user_id)
-        # convert to bytes so the lzma can be deocded with osrparse.
-        replay_data_bytes = base64.b64decode(replay_data_string)
-        parsed_replay = osrparse.parse_replay(replay_data_bytes, pure_lzma=True)
-        replay_data = parsed_replay.play_data
-        if(cache):
-            # Bytes is actually smaller than the b64 encoded string here, so we store that (TODO: compress (well) before storage)
-            Cacher.cache(map_id, user_id, replay_data_bytes)
-        return Replay(replay_data, user_id)
-
-    @staticmethod
-    def from_path(path):
-        """
-        Creates a Replay instance from the data contained by file at the given path.
-
-        Args:
-            [String or Path] path: The absolute path to the replay file.
-
-        Returns:
-            The Replay instance created from the given path.
-        """
-
-        parsed_replay = osrparse.parse_replay_file(path)
-        check_replay_data = parsed_replay.play_data
-        player_name = parsed_replay.player_name
-
-        return Replay(check_replay_data, player_name)
 
     @staticmethod
     def interpolate(data1, data2, interpolation=Interpolation.linear, unflip=False):
